requirements.txt
gunicorn==20.1.0
falcon==3.1.1
opencv-contrib-python-headless==4.7.0.72
opencv-python-headless==4.7.0.72
Pillow==9.5.0
requests==2.30.0
urllib3==2.0.2
